LayoutTests/http/tests/contentextensions/sync-xhr-redirection-blocked-expected.txt

CONSOLE MESSAGE: Content blocker prevented frame displaying http://127.0.0.1:8000/contentextensions/sync-xhr-redirection-blocked.html from loading a resource from http://127.0.0.1:8000/resources/redirect.php?url=/contentextensions/resources/url-blocking-test.js
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8000/resources/redirect.php?url=/contentextensions/resources/url-blocking-test.js.
Synchronous status: 0, readyState:1, responseText:
Synchronous status: 0, readyState:4, responseText:
Synchronous error: NetworkError: A network error occurred.
